removed fdmdv2_ prefix from file names
