+++ src/dlg_plugin.cpp
@@ -21,7 +21,7 @@
 //==========================================================================
 
 #include "dlg_plugin.h"
-#include "fdmdv2_main.h"
+#include "main.h"
 
 #ifdef __WIN32__
 #include <wx/msw/registry.h>
